Addressing first PR comment
